Initial brute force pass on updating language specification for the shift to 64 bit integers (aka Long) as the default integral type for X10 2.4.
Basic approach to the update was to use Long instead of Int in
all examples where possible to do so.  The remaining examples with
Int are making specific points about fixed-size integers and/or
coercion.  All example code snippets compile, but we really need to
make a 'semantic' read over the language specification as well before
the X10 2.4 release.



git-svn-id: svn://svn.code.sf.net/p/x10/code/documentation/trunk@25929 a326200e-df2c-42e0-81fa-597e909af41d